# SWAGGER-DOCS/aula01/src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: API de Usuários
  version: 1.0.0
  description: Documentação da API usando Express e MVC

servers:
  - url: http://localhost:3000

paths:
  /usuarios:
    get:
      summary: Listar usuários
      tags:
        - Usuários
      responses:
        "200":
          description: Usuários encontrados

    post:
      summary: Criar usuário
      tags:
        - Usuários
      requestBody:
        required: true
        content:
          application/json:
            example:
              nome: Joel
              email: joel@email.com
              telefone: (84) 4002-8922
      responses:
        "201":
          description: Usuário criado com sucesso
        "400":
          description: Todos os campos devem ser preenchidos

  /usuarios/{id}:
    get:
      summary: Busca usuário por ID
      tags:
        - Usuários
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        "200":
          description: Usuário encontrado
          content:
            application/json:
              example:
                msg: Usuário encontrado
                usuario:
                  id: 1
                  nome: Amanda
                  email: amanda@email.com
                  telefone: "(84) 9999-9999"
        "404":
          description: Usuário não encontrado

    put:
      summary: Atualizar usuário
      tags:
        - Usuários
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            example:
              nome: Joel
              email: joel@email.com
              telefone: (84) 4002-8922
      responses:
        "200":
          description: Usuário atualizado com sucesso
        "400":
          description: Todos os campos devem ser preenchidos
        "404":
          description: Usuário não encontrado
        "500":
          description: Erro interno ao atualizar usuário

    delete:
      summary: Deletar usuário
      tags:
        - Usuários
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Usuário deletado com sucesso
        "404":
          description: Usuário não encontrado
        "500":
          description: Erro interno ao deletar o usuário